ResizeImageFunction moved to own subnamespace
