+++ tests/ResizeImageFunctionTest.php
@@ -2,12 +2,11 @@
 
 namespace Imponeer\Smarty\Extensions\Image\Tests;
 
+use Imponeer\Smarty\Extensions\Image\Exceptions\AtLeastWidthOrHeightMustBeUsedException;
 use Imponeer\Smarty\Extensions\Image\Exceptions\AttributeMustBeNumericException;
 use Imponeer\Smarty\Extensions\Image\Exceptions\AttributeMustBeStringException;
 use Imponeer\Smarty\Extensions\Image\Exceptions\BadFitValueException;
-use Imponeer\Smarty\Extensions\Image\Exceptions\AtLeastWidthOrHeightMustBeUsedException;
 use Imponeer\Smarty\Extensions\Image\Exceptions\RequiredArgumentException;
-use Imponeer\Smarty\Extensions\Image\ResizeImageFunction;
 use Imponeer\Smarty\Extensions\Image\SmartyImageExtension;
 use Intervention\Image\Exception\NotReadableException;
 use PHPUnit\Framework\Attributes\DataProvider;
